requirements.txt
async-generator==1.10
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
cryptography==36.0.2
h11==0.13.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
lxml==4.5.2
more-itertools==8.4.0
numpy==1.20.1
outcome==1.1.0
packaging==20.4
pandas==1.2.3
progressbar2==3.53.1
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
PyYAML==5.4.1
ratelimiter==1.2.0.post0
requests==2.25.1
selenium==4.1.3
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.4
wsproto==1.1.0
zipp==3.4.1
